+++ datadog/fwprovider/framework_provider.go
@@ -32,6 +32,7 @@ var _ provider.Provider = &FrameworkProvider{}
 
 var Resources = []func() resource.Resource{
 	NewAgentlessScanningAwsScanOptionsResource,
+	NewAgentlessScanningAzureScanOptionsResource,
 	NewAgentlessScanningGcpScanOptionsResource,
 	NewOpenapiApiResource,
 	NewAPIKeyResource,
@@ -50,6 +51,7 @@ var Resources = []func() resource.Resource{
 	NewIntegrationCloudflareAccountResource,
 	NewIntegrationConfluentAccountResource,
 	NewIntegrationConfluentResourceResource,
+	NewIntegrationDatabricksAccountResource,
 	NewIntegrationFastlyAccountResource,
 	NewIntegrationFastlyServiceResource,
 	NewIntegrationGcpResource,
@@ -105,6 +107,9 @@ var Resources = []func() resource.Resource{
 	NewOrgGroupPolicyResource,
 	NewOrgGroupPolicyOverrideResource,
 	NewComplianceResourceEvaluationFilter,
+	NewSecurityMonitoringDefaultRuleResource,
+	NewSecurityMonitoringFilterResource,
+	NewSecurityMonitoringRuleResource,
 	NewSecurityMonitoringRuleJSONResource,
 	NewComplianceCustomFrameworkResource,
 	NewCostBudgetResource,
@@ -608,6 +613,12 @@ func defaultConfigureFunc(p *FrameworkProvider, request *provider.ConfigureReque
 	ddClientConfig.SetUnstableOperationEnabled("v2.CreateDataset", true)
 	ddClientConfig.SetUnstableOperationEnabled("v2.UpdateDataset", true)
 	ddClientConfig.SetUnstableOperationEnabled("v2.DeleteDataset", true)
+
+	ddClientConfig.SetUnstableOperationEnabled("v2.CreateWebIntegrationAccount", true)
+	ddClientConfig.SetUnstableOperationEnabled("v2.GetWebIntegrationAccount", true)
+	ddClientConfig.SetUnstableOperationEnabled("v2.ListWebIntegrationAccounts", true)
+	ddClientConfig.SetUnstableOperationEnabled("v2.UpdateWebIntegrationAccount", true)
+	ddClientConfig.SetUnstableOperationEnabled("v2.DeleteWebIntegrationAccount", true)
 
 	// Enable Logs Restriction Queries
 	ddClientConfig.SetUnstableOperationEnabled("v2.CreateRestrictionQuery", true)