+++ examples/resources/datadog_agentless_scanning_azure_scan_options/import.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+# Import existing agentless scanning Azure scan options by Azure subscription ID
+terraform import datadog_agentless_scanning_azure_scan_options.example 12345678-1234-1234-1234-123456789012